[Lifting State Up, Again] * remove state squares from Board * add state history in Game * move game logic to Game
